* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Text review

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

* Draft

+++ dbpxy/src/main/java/com/dbpxy/service/ThreadFactory.java
@@ -9,9 +9,9 @@ package com.dbpxy.service;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -30,10 +30,10 @@ import org.jspecify.annotations.NonNull;
 class ThreadFactory implements java.util.concurrent.ThreadFactory {
     private final String prefix;
     @Builder.Default
-    private int counter = 0;
+    private long counter = 0L;
 
     @Override
     public Thread newThread(@NonNull final Runnable r) {
-        return new Thread(r, prefix + (counter++)); // ASAP
+        return new Thread(r, prefix + (counter++));
     }
 }
